* Rearranged some constructor initializations to prevent compiler warnings. * Changed powf() to pow() to work around a mingw runtime bug. See http://bugs.debian.org/484299 * Changed abs(float) to fabs(float). abs() only has an int and long version, and mingw was complaining about ambiguity. (float is what we wanted anyway). * Commented out a couple of luabind functions that were preventing mingw from compiling. According to the comments, they gave gcc problems too, and were not used.

+++ demo/src/engine/video/particle_system.cpp
@@ -752,8 +752,8 @@ void ParticleSystem::_UpdateParticles(float t, const EffectParameters &params)
 
 		if(_particles[j].damping != 1.0f)
 		{
-			_particles[j].velocity_x *= powf(_particles[j].damping, t);
-			_particles[j].velocity_y *= powf(_particles[j].damping, t);
+			_particles[j].velocity_x *= pow(_particles[j].damping, t);
+			_particles[j].velocity_y *= pow(_particles[j].damping, t);
 		}
 
 		_particles[j].time += t;
